refactor: move things into their respective modules
