+++ rust/tests/loadtest/src/config.rs
@@ -6,13 +6,10 @@
 /// Minimum ping count (must send at least 1 ping).
 pub const MIN_PING_COUNT: usize = 1;
 
-use rand::SeedableRng;
 use rand::prelude::*;
-use rand::rngs::StdRng;
 use serde::Deserialize;
 use std::net::{IpAddr, SocketAddr};
 use std::path::Path;
-use std::time::Duration;
 use url::Url;
 
 use crate::ping::MAX_ICMP_PAYLOAD_SIZE;
@@ -380,197 +377,6 @@ impl std::fmt::Display for TestType {
             Self::Tcp => write!(f, "tcp"),
             Self::Websocket => write!(f, "websocket"),
             Self::Ping => write!(f, "ping"),
-        }
-    }
-}
-
-/// Resolved HTTP test parameters (ready to execute).
-#[derive(Debug)]
-pub struct ResolvedHttpConfig {
-    pub address: Url,
-    pub http_version: u8,
-    pub users: u64,
-    pub run_time: Duration,
-}
-
-/// Resolved TCP test parameters (ready to execute).
-#[derive(Debug)]
-pub struct ResolvedTcpConfig {
-    pub address: SocketAddr,
-    pub concurrent: usize,
-    pub duration: Duration,
-    pub timeout: Duration,
-    pub echo_mode: bool,
-    pub echo_payload_size: usize,
-    pub echo_interval: Option<Duration>,
-    pub echo_read_timeout: Duration,
-}
-
-/// Resolved WebSocket test parameters (ready to execute).
-#[derive(Debug)]
-pub struct ResolvedWebsocketConfig {
-    pub address: Url,
-    pub concurrent: usize,
-    pub duration: Duration,
-    pub timeout: Duration,
-    pub ping_interval: Option<Duration>,
-    pub echo_mode: bool,
-    pub echo_payload_size: usize,
-    pub echo_interval: Option<Duration>,
-    pub echo_read_timeout: Duration,
-}
-
-/// Resolved ping test parameters (ready to execute).
-#[derive(Debug)]
-pub struct ResolvedPingConfig {
-    /// Target IP addresses.
-    pub targets: Vec<IpAddr>,
-    /// Number of pings per target.
-    pub count: usize,
-    /// Interval between pings.
-    pub interval: Duration,
-    /// Ping timeout.
-    pub timeout: Duration,
-    /// Payload size.
-    pub payload_size: usize,
-}
-
-/// Resolved test configuration (one of the test types).
-#[derive(Debug)]
-pub enum ResolvedConfig {
-    Http(ResolvedHttpConfig),
-    Tcp(ResolvedTcpConfig),
-    Websocket(ResolvedWebsocketConfig),
-    Ping(ResolvedPingConfig),
-}
-
-/// Random test selector.
-pub struct TestSelector {
-    rng: StdRng,
-    seed: u64,
-}
-
-impl TestSelector {
-    /// Create a new selector with the given seed, or generate a random one.
-    pub fn new(seed: Option<u64>) -> Self {
-        let seed = seed.unwrap_or_else(rand::random);
-        let rng = StdRng::seed_from_u64(seed);
-        Self { rng, seed }
-    }
-
-    pub fn seed(&self) -> u64 {
-        self.seed
-    }
-
-    pub fn select(&mut self, config: &LoadTestConfig) -> ResolvedConfig {
-        let types = config.enabled_types();
-        let test_type = types[self.rng.gen_range(0..types.len())];
-
-        match test_type {
-            TestType::Http => ResolvedConfig::Http(self.resolve_http(&config.http)),
-            TestType::Tcp => ResolvedConfig::Tcp(self.resolve_tcp(&config.tcp)),
-            TestType::Websocket => {
-                ResolvedConfig::Websocket(self.resolve_websocket(&config.websocket))
-            }
-            TestType::Ping => ResolvedConfig::Ping(self.resolve_ping(&config.ping)),
-        }
-    }
-
-    fn resolve_http(&mut self, config: &HttpConfig) -> ResolvedHttpConfig {
-        let addr_str = &config.addresses[self.rng.gen_range(0..config.addresses.len())];
-        let address = Url::parse(addr_str).expect("URL validated during config load");
-        let http_version = config.http_version[self.rng.gen_range(0..config.http_version.len())];
-        let users = config.users.pick(&mut self.rng);
-        let run_time = Duration::from_secs(config.run_time_secs.pick(&mut self.rng));
-
-        ResolvedHttpConfig {
-            address,
-            http_version,
-            users,
-            run_time,
-        }
-    }
-
-    fn resolve_tcp(&mut self, config: &TcpConfig) -> ResolvedTcpConfig {
-        let addr_str = &config.addresses[self.rng.gen_range(0..config.addresses.len())];
-        let address: SocketAddr = addr_str
-            .parse()
-            .expect("Address validated during config load");
-
-        let concurrent = config.concurrent.pick(&mut self.rng) as usize;
-        let duration = Duration::from_secs(config.duration_secs.pick(&mut self.rng));
-        let timeout = Duration::from_secs(config.timeout_secs.pick(&mut self.rng));
-        let echo_mode = config.echo_mode;
-        let echo_payload_size = config.echo_payload_size.pick(&mut self.rng) as usize;
-        let echo_interval = Some(Duration::from_secs(
-            config.echo_interval_secs.pick(&mut self.rng),
-        ));
-        let echo_read_timeout =
-            Duration::from_secs(config.echo_read_timeout_secs.pick(&mut self.rng));
-
-        ResolvedTcpConfig {
-            address,
-            concurrent,
-            duration,
-            timeout,
-            echo_mode,
-            echo_payload_size,
-            echo_interval,
-            echo_read_timeout,
-        }
-    }
-
-    fn resolve_websocket(&mut self, config: &WebsocketConfig) -> ResolvedWebsocketConfig {
-        let addr_str = &config.addresses[self.rng.gen_range(0..config.addresses.len())];
-        let address = Url::parse(addr_str).expect("URL validated during config load");
-
-        let concurrent = config.concurrent.pick(&mut self.rng) as usize;
-        let duration = Duration::from_secs(config.duration_secs.pick(&mut self.rng));
-        let timeout = Duration::from_secs(config.timeout_secs.pick(&mut self.rng));
-        let ping_interval = Some(Duration::from_secs(
-            config.ping_interval_secs.pick(&mut self.rng),
-        ));
-        let echo_mode = config.echo_mode;
-        let echo_payload_size = config.echo_payload_size.pick(&mut self.rng) as usize;
-        let echo_interval = Some(Duration::from_secs(
-            config.echo_interval_secs.pick(&mut self.rng),
-        ));
-        let echo_read_timeout =
-            Duration::from_secs(config.echo_read_timeout_secs.pick(&mut self.rng));
-
-        ResolvedWebsocketConfig {
-            address,
-            concurrent,
-            duration,
-            timeout,
-            ping_interval,
-            echo_mode,
-            echo_payload_size,
-            echo_interval,
-            echo_read_timeout,
-        }
-    }
-
-    fn resolve_ping(&mut self, config: &PingConfig) -> ResolvedPingConfig {
-        // Parse all targets
-        let targets: Vec<IpAddr> = config
-            .addresses
-            .iter()
-            .map(|s| s.parse().expect("IP address validated during config load"))
-            .collect();
-
-        // Ensure minimum count of 1 ping
-        let count = (config.count.pick(&mut self.rng) as usize).max(MIN_PING_COUNT);
-        let interval = Duration::from_millis(config.interval_ms.pick(&mut self.rng));
-        let timeout = Duration::from_millis(config.timeout_ms.pick(&mut self.rng));
-        let payload_size = config.payload_size.pick(&mut self.rng) as usize;
-
-        ResolvedPingConfig {
-            targets,
-            count,
-            interval,
-            timeout,
-            payload_size,
         }
     }
 }
